refactor: Storage interface update to support encrypted data
As per the issue #26, the storage shouldn't be aware of the data for privacy reasons. Updated the interface to support this design.

RFC https://github.com/hyperledger/aries-rfcs/blob/master/concepts/0050-wallets/README.md#composition
